move tree_conj, tree_dot, tree_cast, tree_axpy to nkjax
