+++ netket/jax/__init__.py
@@ -23,6 +23,10 @@ from .utils import (
     dtype_complex,
     dtype_real,
     maybe_promote_to_complex,
+    tree_conj,
+    tree_dot,
+    tree_cast,
+    tree_axpy,
     HashablePartial,
     mpi_split,
     PRNGKey,
